fix: 공통 스타일 빼기
