+++ app/domain/signup/components/SignupFunnel/SetName/index.tsx
@@ -1,6 +1,7 @@
 import { InputBox } from '@/common/components/InputBox';
 import { useSignupError, useSignupForm } from '@/domain/signup/stores/SignupFormStore';
 
+import * as c from '../style.css';
 import * as s from './style.css';
 
 const SetName = () => {
@@ -9,14 +10,14 @@ const SetName = () => {
   const { error, clearError } = useSignupError();
 
   return (
-    <div className={s.Wrapper}>
-      <div className={s.Guide}>
-        <p className={s.Description}>나중에 수정 가능해요!</p>
+    <div className={c.Wrapper}>
+      <div className={c.Guide}>
+        <p className={c.Description}>나중에 수정 가능해요!</p>
         <h2>
-          <strong className={s.Strong}>닉네임</strong>을 입력해주세요.
+          <strong className={c.Strong}>닉네임</strong>을 입력해주세요.
         </h2>
       </div>
-      <div className={s.FormWrapper}>
+      <div className={c.FormWrapper}>
         <InputBox
           placeholder="닉네임"
           value={name}
@@ -27,7 +28,7 @@ const SetName = () => {
         />
         <div className={s.InputStatus}>
           <span>{name.length}/10</span>
-          {error === 'name' && <span className={s.ErrorMessage}>이미 존재하는 닉네임 입니다.</span>}
+          {error === 'name' && <span className={c.ErrorMessage}>이미 존재하는 닉네임 입니다.</span>}
         </div>
       </div>
     </div>
